+++ Pandora/Paradigm/Controlflow/Joint.hs
@@ -1,6 +1,6 @@
 module Pandora.Paradigm.Controlflow.Joint (module Exports) where
 
-import Pandora.Paradigm.Controlflow.Joint.Schemes as Exports
+import Pandora.Paradigm.Schemes as Exports
 import Pandora.Paradigm.Controlflow.Joint.Schematic as Exports
 import Pandora.Paradigm.Controlflow.Joint.Transformer as Exports
 import Pandora.Paradigm.Controlflow.Joint.Interpreted as Exports