+++ Pandora/Paradigm/Inventory/Accumulator.hs
@@ -16,7 +16,7 @@ import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, run
 import Pandora.Paradigm.Controlflow.Joint.Transformer.Monadic (Monadic (lay, wrap), (:>) (TM))
 import Pandora.Paradigm.Controlflow.Joint.Schematic (Schematic)
 import Pandora.Paradigm.Controlflow.Joint.Adaptable (Adaptable (adapt))
-import Pandora.Paradigm.Controlflow.Joint.Schemes.UT (UT (UT), type (<.:>))
+import Pandora.Paradigm.Schemes.UT (UT (UT), type (<.:>))
 
 newtype Accumulator e a = Accumulator (e :*: a)