+++ Pandora/Paradigm/Inventory/Environment.hs
@@ -15,7 +15,7 @@ import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, run
 import Pandora.Paradigm.Controlflow.Joint.Transformer.Monadic (Monadic (lay, wrap), (:>) (TM))
 import Pandora.Paradigm.Controlflow.Joint.Schematic (Schematic)
 import Pandora.Paradigm.Controlflow.Joint.Adaptable (Adaptable (adapt))
-import Pandora.Paradigm.Controlflow.Joint.Schemes.TU (TU (TU), type (<:.>))
+import Pandora.Paradigm.Schemes.TU (TU (TU), type (<:.>))
 
 newtype Environment e a = Environment (e -> a)