+++ Pandora/Paradigm/Primary/Functor/Conclusion.hs
@@ -20,7 +20,7 @@ import Pandora.Paradigm.Controlflow.Joint.Interpreted (Interpreted (Primary, run
 import Pandora.Paradigm.Controlflow.Joint.Transformer.Monadic (Monadic (lay, wrap), (:>) (TM))
 import Pandora.Paradigm.Controlflow.Joint.Schematic (Schematic)
 import Pandora.Paradigm.Controlflow.Joint.Adaptable (Adaptable (adapt))
-import Pandora.Paradigm.Controlflow.Joint.Schemes.UT (UT (UT))
+import Pandora.Paradigm.Schemes.UT (UT (UT))
 
 data Conclusion e a = Failure e | Success a