+++ CHANGELOG.md
@@ -235,3 +235,4 @@
 * Define `Zipper` type family to walk datastructures with context
 * Define `Tap` datatype for context dependent values
 * Define type synonyms for `TU`, `UT`, `TUT` and `UTU` joint schemes
+* Move `Schemes` module to `Paradigm` submodule